Move Linkity config to Espoo env
